% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rule_strings.R
\name{.build_rule_strings}
\alias{.build_rule_strings}
\title{Title}
\usage{
.build_rule_strings(
  leaf_paths,
  bin_spec,
  max_depth = NULL,
  tighten_monotone = TRUE
)
}
\arguments{
\item{tighten_monotone}{}
}
\description{
Title
}
\keyword{internal}
