% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_leaf_paths.R
\name{.extract_leaf_paths}
\alias{.extract_leaf_paths}
\title{Title}
\usage{
.extract_leaf_paths(native_leaf_ids, Tm, tdt, helpers, trees_per_batch = 250L)
}
\arguments{
\item{trees_per_batch}{}
}
\description{
Title
}
\keyword{internal}
