% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_max_depth.R
\name{.infer_max_depth}
\alias{.infer_max_depth}
\title{Find maximum realized tree depth in an xgboost model.}
\usage{
.infer_max_depth(tdt)
}
\arguments{
\item{tdt}{}
}
\description{
Find maximum realized tree depth in an xgboost model.
}
\keyword{internal}
