% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_leaf_paths.R
\name{.map_prefixes_to_leaves}
\alias{.map_prefixes_to_leaves}
\title{Title  Map desired prefixes -> (Tree, leaf_id) over ALL leaves.}
\usage{
.map_prefixes_to_leaves(
  want_prefix,
  leaf_paths,
  harvest_bins,
  max_depth,
  tighten_monotone,
  proper_only = TRUE
)
}
\arguments{
\item{want_prefix}{data.table/data.frame with columns:}

\item{proper_only}{}
}
\value{
data.table(prefix_len, rule_str_prefix, Tree, leaf_id) keyed by (prefix_len, rule_str_prefix)
}
\description{
Title  Map desired prefixes -> (Tree, leaf_id) over ALL leaves.
}
\keyword{internal}
