% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_rules.R
\name{harvest_rules}
\alias{harvest_rules}
\title{Title}
\usage{
harvest_rules(
  boosted,
  min_support = 20L,
  trees_per_batch = 250L,
  max_depth = NULL,
  progress_every = NULL,
  trees_subset = NULL,
  tighten_monotone = TRUE
)
}
\arguments{
\item{tighten_monotone}{}
}
\description{
Title
}
