% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_null.R
\name{permute_null}
\alias{permute_null}
\title{Title}
\usage{
permute_null(
  boosted,
  gain_grid = seq(0.05, 0.5, 0.05),
  R = 1000L,
  alpha = 0.5,
  method = c("hard_label", "soft_label"),
  model_llr_weight = 0.5,
  progress_every = NULL
)
}
\arguments{
\item{progress_every}{}
}
\description{
Title
}
