% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_rule_depth.R
\name{analyze_rule_depth}
\alias{analyze_rule_depth}
\title{Title}
\usage{
analyze_rule_depth(
  boosted,
  harvest,
  candidate_rules,
  which = c("train", "test"),
  shrink_m = 0,
  alpha = 0.5,
  fold_indices = NULL,
  progress_every = NULL
)
}
\arguments{
\item{progress_every}{}
}
\description{
Title
}
