% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_rule_depth.R
\name{analyze_rule_depth}
\alias{analyze_rule_depth}
\title{Title}
\usage{
analyze_rule_depth(
  boosted,
  harvest,
  max_depth = NULL,
  top_n = 1000L,
  progress_every = NULL,
  alpha = 0.5
)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
