% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_sensitivity.R
\name{background_sensitivity}
\alias{background_sensitivity}
\title{Title}
\usage{
background_sensitivity(
  boosted,
  extreme_k = 1.51,
  lower_tail = FALSE,
  center = c("mean", "median"),
  scale = c("sd", "mad"),
  k_grid = seq(0.5, 1.5, 0.25),
  R = 1000L,
  alpha = 0.5,
  gain_grid = seq(0.05, 0.5, 0.05),
  verbose = FALSE,
  progress_every = NULL
)
}
\arguments{
\item{progress_every}{}
}
\description{
Title
}
