% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_leaf_matrix.R
\name{.build_leaf_matrix}
\alias{.build_leaf_matrix}
\title{Sparse leaf membership matrix for permutation tests}
\usage{
.build_leaf_matrix(train_leaf_map, N_index_train)
}
\arguments{
\item{N_index_train}{}
}
\description{
Sparse leaf membership matrix for permutation tests
}
\keyword{internal}
