% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{.build_lookup}
\alias{.build_lookup}
\title{Title}
\usage{
.build_lookup(dense_leaf_ids, native_leaf_ids, extr_idx, bg_idx, all = FALSE)
}
\arguments{
\item{all}{}
}
\description{
Title
}
\keyword{internal}
