% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{.build_lookup}
\alias{.build_lookup}
\title{Internal dispatcher for build_lookup}
\usage{
.build_lookup(
  dense_leaf_ids,
  native_leaf_ids,
  extr_idx = NULL,
  bg_idx = NULL,
  all = NULL,
  use_rcpp = TRUE
)
}
\description{
Internal dispatcher for build_lookup
}
\keyword{internal}
