% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_test_leaf_map.R
\name{.build_test_leaf_map}
\alias{.build_test_leaf_map}
\title{Title}
\usage{
.build_test_leaf_map(test_leaves, train_leaf_map)
}
\arguments{
\item{train_leaf_map}{}
}
\description{
Title
}
\keyword{internal}
