% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacc.R
\name{.drop_all_na_cols}
\alias{.drop_all_na_cols}
\title{Drop entirely NA columns from data table}
\usage{
.drop_all_na_cols(DT)
}
\arguments{
\item{DT}{}
}
\description{
Drop entirely NA columns from data table
}
\keyword{internal}
