% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_llrs.R
\name{.leaf_llrs}
\alias{.leaf_llrs}
\title{Title}
\usage{
.leaf_llrs(
  extr_idx,
  bg_idx,
  train_leaf_map,
  N_extr,
  N_bg,
  tree_idx,
  alpha = 0.5,
  return_ids = FALSE
)
}
\arguments{
\item{return_ids}{}
}
\description{
Title
}
\keyword{internal}
