% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_lookup.R
\name{.snp_lookup}
\alias{.snp_lookup}
\title{Internal dispatcher for snp_lookup}
\usage{
.snp_lookup(
  pairs,
  snps_ext_by_leaf = NULL,
  snps_bg_by_leaf = NULL,
  snps_all_by_leaf = NULL,
  use_rcpp = TRUE
)
}
\description{
Internal dispatcher for snp_lookup
}
\keyword{internal}
