% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_lookup.R
\name{.snp_lookup_R}
\alias{.snp_lookup_R}
\title{Title}
\usage{
.snp_lookup_R(
  pairs,
  snps_ext_by_leaf,
  snps_bg_by_leaf,
  snps_all_by_leaf = NULL
)
}
\arguments{
\item{snps_all_by_leaf}{}
}
\description{
Title
}
\keyword{internal}
