% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_indices.R
\name{.validate_indices}
\alias{.validate_indices}
\title{Title}
\usage{
.validate_indices(
  n_yvar_train,
  extr_idx_train,
  bg_idx_train,
  n_yvar_test,
  extr_idx_test,
  bg_idx_test,
  caller = ".validate_indices"
)
}
\arguments{
\item{caller}{}
}
\description{
Title
}
\keyword{internal}
