% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_yvar.R
\name{.validate_yvar}
\alias{.validate_yvar}
\title{Check a y vector for compatibility with an xgboost model and return center and scale}
\usage{
.validate_yvar(y, features, caller = ".validate_yvar")
}
\arguments{
\item{y}{vector of response values used either to train or test an xgboost model}

\item{features}{matrix of feature data}

\item{caller}{an optional string to customize error messages}
}
\description{
Checks if a y-variable vector is 1) numeric with finite center, 2) has finite and positive scale, and 3) is the length expected from the feature matrix. If valid, returns the mean, median, sd, and mad of the vector. Called by \code{make_boosted}.
}
\keyword{internal}
