% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_sensitivity.R
\name{extreme_sensitivity}
\alias{extreme_sensitivity}
\title{Title}
\usage{
extreme_sensitivity(
  boosted,
  background_k = 1,
  lower_tail = FALSE,
  center = c("mean", "median"),
  scale = c("sd", "mad"),
  k_grid = seq(1.25, 1.5, 0.25),
  R = 1000L,
  alpha = 0.5,
  gain_grid = seq(0.05, 0.5, 0.05),
  verbose = FALSE,
  progress_every = NULL
)
}
\arguments{
\item{progress_every}{}
}
\description{
Title
}
