% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_rules.R
\name{harvest_rules}
\alias{harvest_rules}
\title{Title}
\usage{
harvest_rules(
  boosted,
  max_depth = 5L,
  min_support = 20L,
  trees_subset = NULL,
  trees_per_batch = 250L,
  progress_every = NULL,
  tighten_monotone = TRUE,
  return_ledger = TRUE
)
}
\arguments{
\item{return_ledger}{}
}
\description{
Title
}
