% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sensitivity.R
\name{label_sensitivity}
\alias{label_sensitivity}
\title{Title}
\usage{
label_sensitivity(
  boosted,
  noise_grid = seq(0, 0.5, 0.1),
  R = 1000L,
  alpha = 0.5,
  gain_grid = seq(0.05, 0.5, 0.05),
  target = c("extreme", "background"),
  progress_every = NULL
)
}
\arguments{
\item{progress_every}{}
}
\description{
Title
}
