% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_boosted.R
\name{make_boosted}
\alias{make_boosted}
\title{Title}
\usage{
make_boosted(
  model,
  features_train,
  features_test,
  yvar_train,
  yvar_test,
  extr_idx_train,
  bg_idx_train,
  extr_idx_test,
  bg_idx_test,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
