% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_harvest.R
\name{prepare_harvest}
\alias{prepare_harvest}
\title{Title}
\usage{
prepare_harvest(
  boosted,
  target_bins = 10L,
  min_per_bin = 50L,
  winsor_prob = 0.01,
  method = c("fd", "quantile")
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
