% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_snp_class.R
\name{set_snp_class}
\alias{set_snp_class}
\title{Title}
\usage{
set_snp_class(
  yvar_train,
  extreme_k,
  lower_tail = FALSE,
  bg_band_k,
  center = c("mean", "median"),
  scale = c("sd", "mad")
)
}
\arguments{
\item{scale}{}

\item{y}{}
}
\description{
Title
}
\examples{
Title

}
