% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_sensitivity.R
\name{tree_sensitivity}
\alias{tree_sensitivity}
\title{Title}
\usage{
tree_sensitivity(
  boosted,
  subsample_grid = seq(0.1, 0.5, 0.1),
  R = 50L,
  alpha = 0.5,
  gain_grid = seq(0.05, 0.5, 0.05),
  topk_frac = 0.05,
  verbose = FALSE,
  progress_every = NULL
)
}
\arguments{
\item{progress_every}{}
}
\description{
Title
}
