% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_rules.R
\name{validate_rules}
\alias{validate_rules}
\title{Title}
\usage{
validate_rules(
  boosted,
  harvest,
  which = c("train", "test"),
  min_support = 1L,
  shrink_m = 0,
  candidate_rules = NULL,
  fold_indices = NULL,
  compute_pq = NULL,
  progress_every = NULL,
  return_ledger = FALSE
)
}
\arguments{
\item{compute_pq}{}
}
\description{
Title
}
