% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_rules.R
\name{validate_rules}
\alias{validate_rules}
\title{Title}
\usage{
validate_rules(
  boosted,
  harvest,
  candidate_rules = NULL,
  min_support = 1L,
  progress_every = NULL,
  return_ledger = TRUE
)
}
\arguments{
\item{return_ledger}{}
}
\description{
Title
}
